NEXORA - Copy/EmailToTicketAI/requirements.txt
python-dotenv
imapclient
mail-parser
requests
langdetect
